Decouple planner from mission planner
